# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/edge/account": {
            "get": {
                "description": "取得帳戶資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帳戶列表"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nodeattr.EdgeAuth"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/accountrouter.errMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "編輯帳戶列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "帳戶列表"
                ],
                "parameters": [
                    {
                        "description": "帳戶列表資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/accountrouter.editAccountData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/accountrouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/accountrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/login": {
            "get": {
                "description": "登入用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登入、登出功能"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "帳號",
                        "name": "account",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密碼(SHA512格式)",
                        "name": "password",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginrouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/loginrouter.errMsg"
                        }
                    }
                }
            }
        },
        "/edge/logout": {
            "get": {
                "description": "登出用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登入、登出功能"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "帳戶",
                        "name": "account",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginrouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/loginrouter.errMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "accountrouter.editAccountData": {
            "type": "object",
            "properties": {
                "editContent": {
                    "$ref": "#/definitions/nodeattr.EdgeAuth"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "accountrouter.errMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "loginrouter.errMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "nodeattr.EdgeAuth": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    }
}